TST: collect IntervalIndex tests
